python_requirements.txt
torch
torchvision
numpy
networkX
tensorboardX
tqdm
jsonlines
Pillow
transformers